feat(notifications): 423 - quality notification refactoring.
