+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/qualitynotification/domain/base/model/QualityNotificationStatus.java
@@ -99,9 +99,6 @@ public enum QualityNotificationStatus {
         return ACTIVE_STATES.contains(this);
     }
 
-    public static List<QualityNotificationStatus> getActiveStates() {
-        return ACTIVE_STATES;
-    }
 
     public static QualityNotificationStatus from(QualityNotificationStatusRequest qualityNotificationStatusRequest) {
         return QualityNotificationStatus.fromStringValue(qualityNotificationStatusRequest.name());
